tests/hetzner: Print output of debootstrap.log.
Very useful to debug failing builds of the rescue image like this one:

http://hydra.nixos.org/build/27061182/nixlog/197

+++ tests/hetzner-backend/rescue-image.nix
@@ -123,7 +123,10 @@ in pkgs.vmTools.runInLinuxImage (pkgs.stdenv.mkDerivation {
     sed -i -e 's/timeout 0/timeout 1/' \
       config/bootloaders/isolinux/isolinux.cfg
 
-    lb build
+    if ! lb build; then
+      cat /build_fake_rescue/chroot/debootstrap/debootstrap.log >&2
+      exit 1
+    fi
 
     kill -TERM $(< repo.pid)
     chmod 0644 live-image-*.iso
